+++ backend/user/delivery/recieveJSON.go
@@ -1,0 +1,21 @@
+package delivery
+
+import (
+	"net/http"
+
+	"github.com/PUArallelepiped/PUN-street-Universal-Access/swagger"
+	"github.com/PUArallelepiped/PUN-street-Universal-Access/user/repository"
+	"github.com/gin-gonic/gin"
+)
+
+func RecieveData(c *gin.Context) {
+	var req swagger.UserData
+	if err := c.ShouldBindJSON(&req); err != nil {
+		c.JSON(http.StatusBadRequest, gin.H{
+			"error": err.Error(),
+		})
+		return
+	}
+	repository.UpdateData(req)
+	c.JSON(http.StatusOK, req)
+}
